move `documents` folder to `course_evaluation`
