--- backend/documents/urls.py
@@ -1,8 +1,0 @@
-"""
-Note: This is intentionally left out.
-Reason: This resource is crossed between the CourseEvaluation and the Document models.
-
-See:
-- `course_evaluations/urls.py`
-- `reviews/urls.py`
-"""
